Split up Release Data Renderer into multiple components
